fix: changes made to pesa2 and dynapesa2
